requirements.txt
# Core LangChain & RAG
langchain==0.2.7
langgraph==0.1.8
langchainhub==0.1.20
langchain-community==0.2.7
langchain-openai==0.1.16
langchain-chroma==0.1.2

# Document Processing
beautifulsoup4==4.12.3
pypdf==3.17.4
python-docx==1.1.0
unstructured==0.11.8

# Vector Store & Embeddings
chromadb==0.4.22

# Web Search (Optional)
tavily-python==0.3.4

# Configuration & Utilities
python-dotenv==1.0.1
pydantic>=2.0.0
requests==2.31.0

# Testing (Optional)
pytest==8.2.2
pytest-asyncio==0.23.5
pytest-cov==4.1.0
pytest-mock==3.12.0
